refactor: creates `makeOrderFixes`
