+++ utils/make-order-and-comments-after-fixes.ts
@@ -3,7 +3,7 @@ import type { TSESLint } from '@typescript-eslint/utils'
 import type { SortingNode } from '../types/sorting-node'
 
 import { makeCommentAfterFixes } from './make-comment-after-fixes'
-import { getNodeRange } from './get-node-range'
+import { makeOrderFixes } from './make-order-fixes'
 
 interface MakeOrderAndCommentsAfterFixesParameters {
   options?: {
@@ -31,62 +31,25 @@ export let makeOrderAndCommentsAfterFixes = ({
   fixer,
   nodes,
 }: MakeOrderAndCommentsAfterFixesParameters): TSESLint.RuleFix[] => {
-  let fixes: TSESLint.RuleFix[] = []
+  let fixes: TSESLint.RuleFix[] = makeOrderFixes({
+    ignoreFirstNodeHighestBlockComment,
+    sortedNodes,
+    sourceCode,
+    options,
+    fixer,
+    nodes,
+  })
 
   for (let max = nodes.length, i = 0; i < max; i++) {
     let sortingNode = nodes.at(i)!
     let sortedSortingNode = sortedNodes.at(i)!
     let { node } = sortingNode
-    let { addSafetySemicolonWhenInline, node: sortedNode } = sortedSortingNode
-    let isNodeFirstNode = node === nodes.at(0)!.node
-    let isSortedNodeFirstNode = sortedNode === nodes.at(0)!.node
+    let { node: sortedNode } = sortedSortingNode
 
     if (node === sortedNode) {
       continue
     }
 
-    let sortedNodeCode = sourceCode.text.slice(
-      ...getNodeRange({
-        ignoreHighestBlockComment:
-          ignoreFirstNodeHighestBlockComment && isSortedNodeFirstNode,
-        node: sortedNode,
-        sourceCode,
-        options,
-      }),
-    )
-    let sortedNodeText = sourceCode.getText(sortedNode)
-    let tokensAfter = sourceCode.getTokensAfter(node, {
-      includeComments: false,
-      count: 1,
-    })
-    let nextToken = tokensAfter.at(0)
-
-    let sortedNextNodeEndsWithSafeCharacter =
-      sortedNodeText.endsWith(';') || sortedNodeText.endsWith(',')
-    let isNextTokenOnSameLineAsNode =
-      nextToken?.loc.start.line === node.loc.end.line
-    let isNextTokenSafeCharacter =
-      nextToken?.value === ';' || nextToken?.value === ','
-    if (
-      addSafetySemicolonWhenInline &&
-      isNextTokenOnSameLineAsNode &&
-      !sortedNextNodeEndsWithSafeCharacter &&
-      !isNextTokenSafeCharacter
-    ) {
-      sortedNodeCode += ';'
-    }
-    fixes.push(
-      fixer.replaceTextRange(
-        getNodeRange({
-          ignoreHighestBlockComment:
-            ignoreFirstNodeHighestBlockComment && isNodeFirstNode,
-          sourceCode,
-          options,
-          node,
-        }),
-        sortedNodeCode,
-      ),
-    )
     fixes = [
       ...fixes,
       ...makeCommentAfterFixes({
